+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2024-02-24
 
 ### Added
-- Initial release of react-native-electron boilerplate
+- Initial release of react-native-electron-platform boilerplate
 - Electron main process with auto-updater
 - Preload script for secure IPC communication
 - Webpack configuration helper for React Native web builds